create JobSummary component
